+++ src/main/java/edu/ie3/simona/api/data/results/ExtResultDataConnection.java
@@ -8,6 +8,7 @@ package edu.ie3.simona.api.data.results;
 
 import edu.ie3.datamodel.models.result.ModelResultEntity;
 import edu.ie3.datamodel.models.result.NodeResult;
+import edu.ie3.datamodel.models.result.system.FlexOptionsResult;
 import edu.ie3.datamodel.models.result.system.SystemParticipantResult;
 import edu.ie3.simona.api.data.ExtOutputDataConnection;
 import edu.ie3.simona.api.data.ontology.ScheduleDataServiceMessage;
@@ -20,6 +21,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.LinkedBlockingQueue;
+import java.util.stream.Stream;
 import org.apache.pekko.actor.ActorRef;
 
 /** Enables data connection of results between SIMONA and SimonaAPI */
@@ -44,10 +46,16 @@ public class ExtResultDataConnection implements ExtOutputDataConnection {
   /** Map uuid to external id of system participants */
   private final Map<UUID, String> participantResultAssetMapping;
 
+  /** Map uuid to external id of participant flex options */
+  private final Map<UUID, String> flexOptionsMapping;
+
   public ExtResultDataConnection(
-      Map<UUID, String> participantResultAssetMapping, Map<UUID, String> gridResultAssetMapping) {
+      Map<UUID, String> participantResultAssetMapping,
+      Map<UUID, String> gridResultAssetMapping,
+      Map<UUID, String> flexOptionsMapping) {
     this.participantResultAssetMapping = participantResultAssetMapping;
     this.gridResultAssetMapping = gridResultAssetMapping;
+    this.flexOptionsMapping = flexOptionsMapping;
   }
 
   /**
@@ -73,9 +81,35 @@ public class ExtResultDataConnection implements ExtOutputDataConnection {
     return participantResultAssetMapping.keySet().stream().toList();
   }
 
+  public List<UUID> getFlexOptionAssets() {
+    return flexOptionsMapping.keySet().stream().toList();
+  }
+
   /** Method that an external simulation can request results from SIMONA as a list. */
   private List<ModelResultEntity> requestResultList(long tick) throws InterruptedException {
-    sendExtMsg(new RequestResultEntities(tick));
+    List<UUID> allExtEntities =
+        Stream.concat(
+                Stream.concat(getFlexOptionAssets().stream(), getGridResultDataAssets().stream()),
+                getParticipantResultDataAssets().stream())
+            .toList();
+    sendExtMsg(new RequestResultEntities(tick, allExtEntities));
+    return receiveWithType(ProvideResultEntities.class).results();
+  }
+
+  private List<ModelResultEntity> requestFlexOptionResultsList(long tick)
+      throws InterruptedException {
+    sendExtMsg(new RequestResultEntities(tick, getFlexOptionAssets()));
+    return receiveWithType(ProvideResultEntities.class).results();
+  }
+
+  private List<ModelResultEntity> requestGridResultsList(long tick) throws InterruptedException {
+    sendExtMsg(new RequestResultEntities(tick, getGridResultDataAssets()));
+    return receiveWithType(ProvideResultEntities.class).results();
+  }
+
+  private List<ModelResultEntity> requestParticiapntResultsList(long tick)
+      throws InterruptedException {
+    sendExtMsg(new RequestResultEntities(tick, getParticipantResultDataAssets()));
     return receiveWithType(ProvideResultEntities.class).results();
   }
 
@@ -84,6 +118,20 @@ public class ExtResultDataConnection implements ExtOutputDataConnection {
    */
   public Map<String, ModelResultEntity> requestResults(long tick) throws InterruptedException {
     return createResultMap(requestResultList(tick));
+  }
+
+  public Map<String, ModelResultEntity> requestFlexOptionResults(long tick)
+      throws InterruptedException {
+    return createResultMap(requestFlexOptionResultsList(tick));
+  }
+
+  public Map<String, ModelResultEntity> requestGridResults(long tick) throws InterruptedException {
+    return createResultMap(requestGridResultsList(tick));
+  }
+
+  public Map<String, ModelResultEntity> requestParticipantResults(long tick)
+      throws InterruptedException {
+    return createResultMap(requestParticiapntResultsList(tick));
   }
 
   protected Map<String, ModelResultEntity> createResultMap(List<ModelResultEntity> results) {
@@ -96,9 +144,12 @@ public class ExtResultDataConnection implements ExtOutputDataConnection {
             resultMap.put(
                 participantResultAssetMapping.get(systemParticipantResult.getInputModel()),
                 systemParticipantResult);
+          } else if (result instanceof FlexOptionsResult flexOptionsResult) {
+            resultMap.put(
+                flexOptionsMapping.get(flexOptionsResult.getInputModel()), flexOptionsResult);
           } else {
             throw new IllegalArgumentException(
-                "ExtResultData can only handle NodeResult's and SystemParticipantResult's!");
+                "ExtResultData can only handle NodeResult's, FlexOptionResult's and SystemParticipantResult's!");
           }
         });
     return resultMap;