+++ src/main/java/edu/ie3/simona/api/simulation/ExtCoSimulation.java
@@ -9,22 +9,26 @@ package edu.ie3.simona.api.simulation;
 import edu.ie3.datamodel.models.result.ModelResultEntity;
 import edu.ie3.datamodel.models.value.Value;
 import edu.ie3.simona.api.data.DataQueueExtSimulationExtSimulator;
+import edu.ie3.simona.api.data.ExtDataConnection;
 import edu.ie3.simona.api.data.ExtInputDataContainer;
 import edu.ie3.simona.api.data.em.ExtEmDataConnection;
 import edu.ie3.simona.api.data.primarydata.ExtPrimaryDataConnection;
 import edu.ie3.simona.api.data.results.ExtResultContainer;
 import edu.ie3.simona.api.data.results.ExtResultDataConnection;
+import edu.ie3.simona.api.exceptions.ExtDataConnectionException;
 import edu.ie3.simona.api.simulation.mapping.DataType;
+import edu.ie3.simona.api.simulation.mapping.ExtEntityEntry;
 import edu.ie3.simona.api.simulation.mapping.ExtEntityMapping;
-import java.util.Map;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
+import java.util.stream.Collectors;
 import org.slf4j.Logger;
 
 /**
  * Abstract class for an external co-simulation with the structure: external api - ext-co-simulation
- * - extsimulation - simonaAPI - simona It contains all function to transfer primary data and em
- * data to SIMONA and results to the external co-simulation.
+ * - ext-simulation - simonaAPI - simona
+ *
+ * <p>It contains all function to transfer primary data and em data to SIMONA and results to the
+ * external co-simulation.
  */
 public abstract class ExtCoSimulation extends ExtSimulation {
 
@@ -46,6 +50,15 @@ public abstract class ExtCoSimulation extends ExtSimulation {
     this.dataQueueSimonaApiToExtCoSimulator = new DataQueueExtSimulationExtSimulator<>();
   }
 
+  @SafeVarargs
+  protected static Set<ExtDataConnection> toSet(
+      Optional<? extends ExtDataConnection>... optionals) {
+    return Arrays.stream(optionals)
+        .filter(Optional::isPresent)
+        .map(Optional::get)
+        .collect(Collectors.toSet());
+  }
+
   /**
    * Builds an {@link ExtPrimaryDataConnection}.
    *
@@ -53,19 +66,32 @@ public abstract class ExtCoSimulation extends ExtSimulation {
    * @param log logger
    * @return an ext primary data connection
    */
-  protected static ExtPrimaryDataConnection buildPrimaryConnection(
+  @SuppressWarnings("unchecked")
+  public static ExtPrimaryDataConnection buildPrimaryConnection(
       ExtEntityMapping mapping, Logger log) {
-    Map<String, UUID> primaryMapping = mapping.getExtId2UuidMapping(DataType.EXT_PRIMARY_INPUT);
-    ExtPrimaryDataConnection extPrimaryDataConnection =
-        new ExtPrimaryDataConnection(primaryMapping);
+    List<ExtEntityEntry> entries = mapping.getExtEntityEntries(DataType.EXT_PRIMARY_INPUT);
 
-    if (primaryMapping.isEmpty()) {
-      log.warn("Primary with 0 entities created.");
+    if (entries.isEmpty()) {
+      log.warn("No primary data connection was created.");
+      throw new ExtDataConnectionException(ExtPrimaryDataConnection.class);
     } else {
-      log.info("Primary connection with {} entities created.", primaryMapping.size());
-    }
+      log.info("Primary data connection with {} entities created.", entries.size());
 
-    return extPrimaryDataConnection;
+      Map<String, UUID> primaryMapping =
+          entries.stream().collect(Collectors.toMap(ExtEntityEntry::id, ExtEntityEntry::uuid));
+      Map<UUID, Class<Value>> valueClasses = new HashMap<>();
+      entries.stream()
+          .filter(e -> e.columnScheme().isPresent())
+          .forEach(
+              e ->
+                  valueClasses.put(
+                      e.uuid(), (Class<Value>) e.columnScheme().get().getValueClass()));
+
+      ExtPrimaryDataConnection primaryDataConnection = new ExtPrimaryDataConnection(primaryMapping);
+      primaryDataConnection.setValueClasses(valueClasses);
+
+      return primaryDataConnection;
+    }
   }
 
   /**
@@ -75,17 +101,16 @@ public abstract class ExtCoSimulation extends ExtSimulation {
    * @param log logger
    * @return an ext em data connection
    */
-  protected static ExtEmDataConnection buildEmConnection(ExtEntityMapping mapping, Logger log) {
+  public static ExtEmDataConnection buildEmConnection(ExtEntityMapping mapping, Logger log) {
     Map<String, UUID> emMapping = mapping.getExtId2UuidMapping(DataType.EXT_EM_INPUT);
-    ExtEmDataConnection extEmDataConnection = new ExtEmDataConnection(emMapping);
 
     if (emMapping.isEmpty()) {
-      log.warn("Em connection with 0 entities created.");
+      log.warn("No em data connection was created.");
+      throw new ExtDataConnectionException(ExtEmDataConnection.class);
     } else {
-      log.info("Em connection with {} entities created.", emMapping.size());
+      log.info("Em data connection with {} entities created.", emMapping.size());
+      return new ExtEmDataConnection(emMapping);
     }
-
-    return extEmDataConnection;
   }
 
   /**
@@ -95,28 +120,55 @@ public abstract class ExtCoSimulation extends ExtSimulation {
    * @param log logger
    * @return an ext result data connection
    */
-  protected static ExtResultDataConnection buildResultConnection(
+  public static ExtResultDataConnection buildResultConnection(
       ExtEntityMapping mapping, Logger log) {
     Map<UUID, String> resultParticipantMapping =
         mapping.getExtUuid2IdMapping(DataType.EXT_PARTICIPANT_RESULT);
     Map<UUID, String> resultGridMapping = mapping.getExtUuid2IdMapping(DataType.EXT_GRID_RESULT);
-    ExtResultDataConnection extResultDataConnection =
-        new ExtResultDataConnection(resultParticipantMapping, resultGridMapping);
+    Map<UUID, String> resultFlexOptionsMapping =
+        mapping.getExtUuid2IdMapping(DataType.EXT_FLEX_OPTIONS_RESULT);
 
-    if (resultParticipantMapping.isEmpty() && resultGridMapping.isEmpty()) {
-      log.warn("Result connection with 0 participants and 0 grid assets created.");
+    if (resultParticipantMapping.isEmpty()
+        && resultGridMapping.isEmpty()
+        && resultFlexOptionsMapping.isEmpty()) {
+      log.warn("No result connection was created.");
+      throw new ExtDataConnectionException(ExtResultDataConnection.class);
     } else {
       log.info(
-          "Result connection with {} participants and {} grid assets created.",
+          "Result connection with {} participants, {} grid assets and {} flex option mappings created.",
           resultParticipantMapping.size(),
-          resultGridMapping.size());
+          resultGridMapping.size(),
+          resultFlexOptionsMapping.size());
+      return new ExtResultDataConnection(
+          resultParticipantMapping, resultGridMapping, resultFlexOptionsMapping);
     }
-
-    return extResultDataConnection;
   }
 
   /**
-   * Function to send primary data to SIMONA using ExtPrimaryData
+   * Function to send primary data to SIMONA using the given {@link ExtPrimaryDataConnection}. This
+   * method will take a value from the {@link #dataQueueExtCoSimulatorToSimonaApi}.
+   *
+   * @param extPrimaryDataConnection the connection to SIMONA
+   * @param tick for which data is sent
+   * @param maybeNextTick option for the next tick data is sent
+   * @param log logger
+   */
+  protected void sendPrimaryDataToSimona(
+      ExtPrimaryDataConnection extPrimaryDataConnection,
+      long tick,
+      Optional<Long> maybeNextTick,
+      Logger log)
+      throws InterruptedException {
+    sendPrimaryDataToSimona(
+        extPrimaryDataConnection,
+        tick,
+        dataQueueExtCoSimulatorToSimonaApi.takeData().getSimonaInputMap(),
+        maybeNextTick,
+        log);
+  }
+
+  /**
+   * Function to send primary data to SIMONA using the given {@link ExtPrimaryDataConnection}.
    *
    * @param extPrimaryDataConnection the connection to SIMONA
    * @param tick for which data is sent
@@ -137,8 +189,33 @@ public abstract class ExtCoSimulation extends ExtSimulation {
   }
 
   /**
-   * Function to send em data to SIMONA using ExtPrimaryData nextTick is necessary, because the em
-   * agents have an own scheduler that should know, when the next set point arrives.
+   * Function to send em data to SIMONA using the given {@link ExtEmDataConnection}. This method
+   * will take a value from the * {@link #dataQueueExtCoSimulatorToSimonaApi}.
+   *
+   * <p>{@code nextTick} is necessary, because the em agents have an own scheduler that should know,
+   * when the next set point arrives.
+   *
+   * @param extEmDataConnection the connection to SIMONA
+   * @param tick for which data is sent
+   * @param maybeNextTick option for the next tick data is sent
+   * @param log logger
+   */
+  protected void sendEmDataToSimona(
+      ExtEmDataConnection extEmDataConnection, long tick, Optional<Long> maybeNextTick, Logger log)
+      throws InterruptedException {
+    sendEmDataToSimona(
+        extEmDataConnection,
+        tick,
+        dataQueueExtCoSimulatorToSimonaApi.takeData().getSimonaInputMap(),
+        maybeNextTick,
+        log);
+  }
+
+  /**
+   * Function to send em data to SIMONA using the given {@link ExtEmDataConnection}.
+   *
+   * <p>{@code nextTick} is necessary, because the em agents have an own scheduler that should know,
+   * when the next set point arrives.
    *
    * @param extEmDataConnection the connection to SIMONA
    * @param tick for which data is sent
@@ -158,14 +235,63 @@ public abstract class ExtCoSimulation extends ExtSimulation {
   }
 
   /**
-   * Function to get result data from SIMONA using the available {@link ExtResultDataConnection}
+   * Function to send only participant result data from SIMONA to the external simulation using the
+   * given {@link ExtResultDataConnection}
+   */
+  protected void sendParticipantResultsToExt(
+      ExtResultDataConnection connection, long tick, Optional<Long> nextTick, Logger log)
+      throws InterruptedException {
+    sendSingleResultType(
+        "participant", connection.requestParticipantResults(tick), tick, nextTick, log);
+  }
+
+  /**
+   * Function to send only grid result data from SIMONA to the external simulation using the given
+   * {@link ExtResultDataConnection}
+   */
+  protected void sendGridResultsToExt(
+      ExtResultDataConnection connection, long tick, Optional<Long> nextTick, Logger log)
+      throws InterruptedException {
+    sendSingleResultType("grid", connection.requestGridResults(tick), tick, nextTick, log);
+  }
+
+  /**
+   * Function to send only flex option result data from SIMONA to the external simulation using the
+   * given {@link ExtResultDataConnection}
+   */
+  protected void sendFlexOptionResultsToExt(
+      ExtResultDataConnection connection, long tick, Optional<Long> nextTick, Logger log)
+      throws InterruptedException {
+    sendSingleResultType(
+        "flex option", connection.requestFlexOptionResults(tick), tick, nextTick, log);
+  }
+
+  private void sendSingleResultType(
+      String type,
+      Map<String, ModelResultEntity> resultsToBeSend,
+      long tick,
+      Optional<Long> nextTick,
+      Logger log)
+      throws InterruptedException {
+    log.info("Request results from SIMONA for {} for tick {}!", type, tick);
+
+    String resultString = resultMapToString(resultsToBeSend);
+    log.debug("[{}] Received {} results from SIMONA!\n{}", tick, type, resultString);
+    dataQueueSimonaApiToExtCoSimulator.queueData(
+        new ExtResultContainer(tick, resultsToBeSend, nextTick));
+    log.info("Sent {} results for tick {} to {}", type, tick, extSimulatorName);
+  }
+
+  /**
+   * Function to send all result data from SIMONA to the external simulation using the given {@link
+   * ExtResultDataConnection}
    *
    * @param connection the connection to SIMONA
    * @param tick for which data is received
    * @param maybeNextTick option for the next tick data is received
    * @param log logger
    */
-  protected void sendDataToExt(
+  protected void sendResultToExt(
       ExtResultDataConnection connection, long tick, Optional<Long> maybeNextTick, Logger log)
       throws InterruptedException {
     log.debug("Request results from SIMONA!");
@@ -174,5 +300,24 @@ public abstract class ExtCoSimulation extends ExtSimulation {
     dataQueueSimonaApiToExtCoSimulator.queueData(
         new ExtResultContainer(tick, resultsToBeSend, maybeNextTick));
     log.debug("Sent results to {}", extSimulatorName);
+  }
+
+  private String resultMapToString(Map<String, ModelResultEntity> results) {
+    StringBuilder resultString = new StringBuilder();
+
+    for (Map.Entry<String, ModelResultEntity> entry : results.entrySet()) {
+      String key = entry.getKey();
+      ModelResultEntity value = entry.getValue();
+
+      resultString
+          .append("id = ")
+          .append(key)
+          .append(", time = ")
+          .append(value.getTime())
+          .append(", result = ")
+          .append(value.getClass().getSimpleName())
+          .append("\n");
+    }
+    return resultString.toString();
   }
 }