+++ src/main/java/edu/ie3/simona/api/simulation/mapping/ExtEntityEntry.java
@@ -8,6 +8,7 @@ package edu.ie3.simona.api.simulation.mapping;
 
 import edu.ie3.datamodel.io.naming.timeseries.ColumnScheme;
 import edu.ie3.datamodel.models.input.InputEntity;
+import java.util.Optional;
 import java.util.UUID;
 
 /**
@@ -15,14 +16,11 @@ import java.util.UUID;
  *
  * @param uuid SIMONA uuid
  * @param id external id
- * @param columnScheme data types the external asset expects
+ * @param columnScheme option for data types the external asset expects
  * @param dataType data types the external asset expects
  */
 public record ExtEntityEntry(
-    UUID uuid,
-    String id,
-    ColumnScheme columnScheme, // FIXME: placeholder -> ColumnScheme should handle more data types
-    DataType dataType)
+    UUID uuid, String id, Optional<ColumnScheme> columnScheme, DataType dataType)
     implements InputEntity {
 
   public String toString() {